Move element operations to right place
